Fix the Change School test
Stop ignoring the test.
In order to account for some delays in loading added some sleep / spin to FeatureContext::accessToSchool to ensure it doesn’t give up too soon.
Use a spinning xpath selector for clicking on tabs.
Wait for permissions to be loaded before attempting to change permissions.
Add FeatureContext:: findXpathElement - makes the behat runs less fragile by searching for elements using path
and dying gracefully when they are not found in a timeout.