Temporarily disable Save All Draft button
While #609 is unresolved don’t ever enable the Save all as Draft
button.  Place some explanatory text in the title to inform users of
why the button is disabled.  Also in case I missed anywhere this button
would be enabled I replaced the action on click with another warning
message.