Rely on username instead of EPPN
Instead of storing the eppn in its own column store it in the username
column of the authentication table.  This opens up a lot of
simplification as we can rely on setting the username from a
configurable LDAP attribute when creating new users.  Then user sync
will work for ldap and shibboleth installations.