Add absoluteFileUri to learning material endpoint
In order to get a path we need access to the symphony router which is
not available in the learning material entity.  So the controller has
to instead use a factory to decorate each entity as it is returned and
processed.  The allows the factory to be a service which receives the
Router as a dependancy.

Also removed token and relativePath from the learning material API as
these are internal implementation items and not needed for public
consumption.