Fail gracefully on missing frontend
Instead of just dying we can be much better with this.  By messaging the
user during cache:clear and then presenting an error page we can ensure
that tests will run even if there is an API mismatch between Ilios and
the fronted and no comparable fronted is yet available.