Reduce expensive calls to Reflection and Annotations
Each time an entity is normalized all of the exposed properties are
enumerated.  This is an expensive process.  We can easily cache this
information for the duration of a single request so when multiple
entities as in getAll?limit=1000 are requested we only run it once.  For
the request courses?limit=10000 this took the response time from 30
seconds to 12 seconds, a substantial improvement.