Intercept requests with a crossing guard
When the crossing guard is down requests will be queued and held until
it is up again. This is useful for putting the API into a very temporary
maintenance mode while running any necessary migrations during
deployment. Instead of blocking and erroring we can simply hold off on
responding to requests until the DB is ready.