Use the DTO pattern for offerings
I have added some non-api properties to the OfferingDTO in this case to
make voter authorization much faster. Instead of looking up course and
school in the manager we can just pull them off of the DTO. Since these
are not annotated for the API they won't show up there.