Use ENV variables for almost all configuration
Everything but forceProtocol (which is still needed in routes) is now
expected to come from the environment.

parameters.yml.dist now file those variables into the environment
instead of providing them to Symfony.

In development these parameters can be provided in a `.env` file in the
root of the project. This shouldn't ever be checked in.