Remove @package annotation everywhere
We're not generating API docs from these blocks, and even if we were the
namespace off each file is a mirror for our `@package` annotation in
every case. Because we use annotation to control the API exposure for
each Entity we actually end up attempting to parse the `@package`
annotation as well, which is a lot of unnecessary requests to
`get_class`.