Fix bad mimetypes in learning materials
We were incorrectly storing some learning materials mimetypes to
link or citation, but that can be fixed.
This command looks through all learning materials to find
ones with a problem and fixes them. First we try and detect the type
from the file itself then we fallback to using the extension on the
filename.