Improve caching of frontend assets
Three different caching strategies are needed:
1) The index.html file should be cached by browsers, but it must be
checked on every request to ensure it is fresh. If it is then return a
304 No Change header
2) sw.js and sw-registration.js are handed like index.html to ensure all
updated get noticed immediately.
3) The rest of our files are fingerprinted and they can be cached
forever (I set it to one year) browsers never need to check back as they
will get a new fingerprint if they are ever updated.