Improve re-indexing to avoid conflicts
When elasticsearch deletes by query it copies the data out and then
attempts to act on it. If another operation takes place on the same data
at the same time it causes a conflict and dies. So instead of deleting
everything in a course index every time I've switched this around to
only delete when a course or session is deleted. This keeps the index up
to date and avoids conflicts since courses are not deleted very often
and sessions can be deleted by ID and not by query which is more
reliable.