Require strict types
As of PHP 7 it is possible to require that type checks only pass on an
exact match and not something that can be coerced into a match. This
declaration is file scoped so it has to be added to every file.

I've addeda linting rule to enforce this declaration on each file and
added the declaration to every file in the project.