Authenticate CAS users immediately
Instead of loading the frontend and then doing redirects this takes an
un-authenticated user directly to the CAS login page, when they are
returned to Ilios they are authenticated and then redirected back to
load index.html. Once the frontend is loaded it will send a request to
authenticate and get back a JWT as it used to.