Store uploaded files in the shared file system
Learning materials are uploaded into a temporary file system on whatever
server the user happens to hit. On distributed systems this makes it a
random chance of catching the system which contains the file when you
want to save it.

This change stores these uploaded files on the official FS (S3, shared
drive) which is already setup to be shared among machines. Thus ensuring
that tmp files are available when users want to attach them to a
learning material.