Handle bulk POST requests with JSON:API
This is aligned with the Bulk data profile which matches closely the bulk creation API we already have. The only difference is that bulk requests return an array for data and single requests return an object. By using the metadata service we can transform existing data into bulk data easily whcih avoids needed to add this boilerplate to every test data factory.