Create Change Alerts When Offerings Are Modified
When an offering is saved or modified a change alert should be created, these were being built, but since the manager wasn't being flushed they were not persisted to the database. The testing infrastructure for this was mostly nutered when the API endpoint for alerts was removed, but by querying the database directly we can once again test to ensure these alerts are working.

For PATH JSON:API requests a new method was added to the Offerings Controller to handle saving these alerts.