Handle LM decoration during serialization
We were doing this work in the controller, but that doesn't account for
side loaded materials. In order to ensure that all materials are
decorated we need to do this work in the serializer instead.

The FactoryNormalizer is built to handle any of our data which has a
factory and to transform it and then pass the result back into the
serializer to continue to be processed.

Then instead of relying on JsonApiDTONormalier directly in the JsonApiData
shaping process we use the full serializer chain which ensures that side
loaded data also passes through the FactoryNormalizer before landing in
the JsonApiDTONormalizer.