Force SSL for CAS redirect
Even if the ilios instance isn't forcing SSL for all connections we
should require it for the CAS service. This ensures that even when
running behind an SSL terminating load balancer the CAS ticket won't be
exposed in the redirect and that there will be a match with the service
URL on the CAS server.