Allow PHP Attributes to mark DTOs
This is the first step in removing our need for dockblock annotations
and the userland libraries needed to read those tags. Initial benchmarks
put this at about 5% faster for the trivial usecase in our tests, it may
be a bigger performance improvement with complicated multi-object
JSON:API requests.

This is POC at this point with only AamcMethodDTO actually making the
transition, but I've modified out lookup to account first for attributes
and then to fallback to the annotations. This should allow a multi-step
transition of all DTOs to this over time.