replace check if data exists with attempt to clear out data.
turns out that when schema migrations are running (an earlier step in
the installation process), some data makes it into some of the
tables that we're trying to populate here.
so we cannot check and block the import if data exists.
therefore, plan B - attempt to scrub those tables first.