Modify nginx to be more tolerant of missing PHP container
When nginx boots it was imemdiatly searching for the php-fpm host which,
if it wasn't present, would then kill nginx. This doesn't work very well
in our docker setup and can intorduce mysterious failures.

This change stores the php-fpm host as an in-nginx-config variable that
isn't read until it is needed so startup isn't blocked. Because nginx
has it's own internal dns resolution that doesn't fall back to the host
we also need to override the resolve of the server. Docker provides DNS
that works well for this, but I've added it as an environmental variable
as well so it can be overridden at boot time if fargate or some other
task runner doesn't something different.