Re-organize how the frontend is stored on disk
Instead of splitting this up into various cache and tmp directories I've
consolidated everything into var/frontend. This is makes it possible to
share this directory with containers making it easier to develop and
test ilios.

I also realized that we were still thinking about the active version as
a path to many assets, but it's really just a single JSON file so I
changed the way we share that path with other areas of the application
and was able to drop some constants and make others private.