Clear up relationships
A programYear has to have a program and a program has to have a school.

An offering has to have a session and a session has to have a course.

A learning material has to have an owning user and that user has to have
a school

These are currently validated before saving so it's safe to assume there
will be no data in the DB that doesn't meet these requirements.

We have allowed cohorts to be created without a program year, but we'll
probably need to clean that up at some point.