Remove invalid comparison check
This could never be true as comparing any value for ID to an object
doesn't make sense. We do allow the CIBlock class to cast itself to a
string by its ID, however since getId() is going to return and int and
we're doing a strict comparison I don't believe even PHP magic would
ever evaluate this to true.