By Specific about DTO list types
In our JSON API we cast all of our relationships to strings. GraphQL is more specific about types and how they relate and allow us to do this anymore so I created the array<integer> type. So as to not change our current API responses we're continuing to cast these as strings there and only using this change in the GraphQL API.