Filter Mesh Descriptor by School
I had to re-structure the way we do joins in the repository so they
would be shared amongst filters otherwise we hit a mysql limit. This
filter set is SLOW, but that's a pre-existing problem with mesh and
adding the schools filter doesn't seem to make it better or worse.