Add a consume-messages-dev container
For development it's super helpful to have the verbose output of
messages being consumed and to include the higher memory and other
symfony development PHP config tweaks. I have the container building,
but not deploying because I don't know if this is something anyone else
would want to use outside of ilios local development.