Share File Storage Between Containers
In order to be able to view uploads and index materials in different
containers we have to share the file storage between them. In production
this is done with S3 usually, but here I've created a names volume and
ensured files are shared there. When the container boots it will create
this directory if it doesn't exist, which is required in order to give
it good permissions, but shouldn't be an issue for any other reason.