Replace elastic with open search
This is sort of a semantic change in that opensearch is currently API
compatible with elastic's PHP client, but we have no idea how long that
will remain true. As Ilios hosting is on AWS and we know other schools
are using that cloud platform as well this switch to opensearch serves
the needs of our users at this time. This was necessary because we are
otherwise unable to update our PHP search dependency and still maintain
compatibility with both systems.