Remove ClockMock and accept the slight delay
This creates a few seconds delay in a half dozen tests, but it removes
an area of flakeyness from our tests when the mocked clock doesn't do
what we expect.

While working on this I made some small changes to the
timestamping application code that I think are keep worthy. I've changed
the variable name and the way we build a DateTime from a timestamp and I
also applied a unique method to the IDs before saving so we don't
duplicate a bunch of IDs when we save things in big groups.